smart_parking/requirements.txt
Flask==3.0.0
mysql-connector-python==8.2.0
Werkzeug==3.0.1
Flask==3.0.0
mysql-connector-python==8.2.0
Werkzeug==3.0.1
Flask-SocketIO==5.3.6
eventlet==0.33.3